day 8 work

+++ backend-server/src/com/project/hrmanagement/Dao/IFeedbackDao.java
@@ -15,11 +15,11 @@ public interface IFeedbackDao {
 		public List<Feedback> listAllFeedback();
 		
 		//search
-		public Feedback searchFeedback(Long feedbackId);
+		public Feedback searchFeedback(Integer feedbackId);
 		
 		
 		//remove
-		public Feedback removeFeedback(Long feedbackId);
+		public Feedback removeFeedback(Integer feedbackId);
 		
 		
 }
